pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-douyin"
version = "0.1.0"
description = "Tools for fetching and processing Douyin product images"
authors = [{name = "Codex"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0.80",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra"
